Moved particle creation routines into a new file
The routines were moved from PhysicalParticleContainer.cpp to
ParticleCreation/AddParticles.cpp, in order to reduce the
size (and compile time) of PhysicalParticleContainer.cpp.
