+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
     - id: check-toml
     - id: check-yaml
     - id: check-added-large-files
-      args: ['--maxkb=40']
+      args: ['--maxkb=80']
     - id: requirements-txt-fixer
 #  - id: fix-encoding-pragma
 #    exclude: ^noxfile.py$
