+++ Source/Particles/CMakeLists.txt
@@ -2,7 +2,6 @@ foreach(D IN LISTS WarpX_DIMS)
     warpx_set_suffix_dims(SD ${D})
     target_sources(lib_${SD}
       PRIVATE
-        AddPlasmaUtilities.cpp
         MultiParticleContainer.cpp
         ParticleBoundaries.cpp
         PhotonParticleContainer.cpp
